backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
pymysql
python-dotenv
bcrypt
requests